UPDATE: api params, sql script

+++ server/SunriseServer.Core/Models/Product.cs
@@ -13,7 +13,7 @@ namespace SunriseServerCore.Models
 
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public byte Discount { get; set; }
+        public double Discount { get; set; }
         public int Fabric { get; set; }
         public string FabricName { get; set; }
         public string Color { get; set; }
